requirements.txt
# =============================================================================
# EmailOps Cortex - Production Dependencies
# =============================================================================
# Minimal set of dependencies for production deployment
# Dev dependencies are in requirements-dev.txt
# =============================================================================

# ---------------------------
# Web Framework
# ---------------------------
fastapi>=0.115.0
uvicorn>=0.30.0
starlette>=0.38.0
python-multipart>=0.0.9

# ---------------------------
# Database & ORM
# ---------------------------
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
alembic>=1.13.0
redis>=5.0.0

# ---------------------------
# LLM & Embeddings
# ---------------------------
openai>=1.50.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.5.0
google-auth>=2.25.0

# ---------------------------
# Document Processing
# ---------------------------
pypdf>=4.0.0
python-docx>=1.0.0
python-pptx>=0.6.23
openpyxl>=3.1.0
extract-msg>=0.50.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ---------------------------
# Observability (OpenTelemetry)
# ---------------------------
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp-proto-http>=1.20.0

# ---------------------------
# Validation & Config
# ---------------------------
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# ---------------------------
# Security & Auth
# ---------------------------
PyJWT[crypto]>=2.8.0
cryptography<44.1.0

# ---------------------------
# HTTP & Networking
# ---------------------------
httpx>=0.26.0
requests>=2.31.0

# ---------------------------
# CLI & Utils
# ---------------------------
typer>=0.12.0
pyyaml>=6.0.0
tenacity>=8.2.0
paramiko>=3.4.0
tiktoken>=0.5.0

# ---------------------------
# Data Processing
# ---------------------------
numpy>=1.24.0
pandas>=2.0.0
mcp==1.23.0
langgraph
